requirements.txt
invoke==0.16.3
Django==3.0
psycopg2-binary==2.8.3
django-redis==4.11.0
celery==4.4.0
redis==3.3.11
uwsgi==2.0.18
google-api-python-client==1.7.11
google-auth==1.10.0
social-auth-app-django==3.1.0
django-celery-results==1.1.2
social-auth-core==3.2.0
oauth2client==4.1.3
hurry.filesize==0.9
ipython==5.3.0
